requirements.txt
Django
djangorestframework
django-filter
psycopg2-binary
gunicorn